Sender email

+++ personal-email/blueprints/customer_group/resources.py
@@ -29,9 +29,11 @@ class CustomerGroupResource(Resource):
     def post(self):
         parser = reqparse.RequestParser()
         parser.add_argument('name', location='json')
+        parser.add_argument('status', location='json')
         args = parser.parse_args()
+        claims = get_jwt_claims()
         
-        customer_group = CustomerGroup( args['name'])
+        customer_group = CustomerGroup( args['name'], claims['id'], args['status'])
 
         db.session.add(customer_group)
         db.session.commit()
@@ -41,11 +43,16 @@ class CustomerGroupResource(Resource):
 
     # delete customer group
     @staff_required
-    def delete(self, id):
+    def patch(self, id):
         qry = CustomerGroup.query.get(id)
-        # if qry is None:
-        #     return {'status': 'NOT_FOUND'}, 404
+
+        qry.status = False
+        if qry is None:
+            return {'status': 'NOT_FOUND'}, 404
+
+     
         db.session.delete(qry)
+
         db.session.commit()
 
 
@@ -59,12 +66,20 @@ class ListCustomerGroup(Resource):
         parser.add_argument('rp', type=int, location='args', default=25)
 
         args = parser.parse_args()
+        claims = get_jwt_claims()
         offset = (args['p']*args['rp']-args['rp'])
+
+        qry = CustomerGroup.query.filter_by(user_id=claims['id'])
+
+        if qry is None:
+            return {'status': 'NOT_FOUND'}, 404
+
         qry = CustomerGroup.query
         print("==============----------------------")
         print(qry)
         # if qry is None:
         #     return {'status': 'NOT_FOUND'}, 404
+
         rows=[]
         for row in qry.offset(offset).all():
             marshal_group= marshal(row, CustomerGroup.response_fields)
